UBUNTU: [Config] oracle: Remove option to lift lockdown via SysRq
BugLink: https://bugs.launchpad.net/bugs/1861238

The original intent behind Lockdown's SysRq support was that the SysRq
command to lift Lockdown would only be honored if the command was
physically entered on a keyboard. Attempts to synthetically generate the
SysRq command, by a software program, were to be ignored since software,
even running as root, must not have the authorization to lift Lockdown.

Unfortunately, attempts to detect a synthetic SysRq command can be
thwarted by a privileged process that is able to set up a USB/IP
connection as the USB/IP connection could be used to lift Lockdown.

Remove the ability to lift Lockdown using SysRq

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>